Optimized and reduced duplicate code
