+++ app/src/main/java/com/itachi1706/cheesecakeutilities/Modules/gpaCalculator/fragment/BaseGpaFragment.kt
@@ -1,12 +1,22 @@
 package com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.fragment
 
 import android.content.Context
+import android.os.Bundle
 import android.util.Log
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.DefaultItemAnimator
+import androidx.recyclerview.widget.ItemTouchHelper
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.google.firebase.database.ValueEventListener
 import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.GpaCalcFirebaseUtils
+import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.GpaRecyclerAdapter
 import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.MainViewActivity
 import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.interfaces.StateSwitchListener
+import com.itachi1706.cheesecakeutilities.R
 import com.itachi1706.cheesecakeutilities.RecyclerAdapters.SwipeEditDeleteCallback
 import com.itachi1706.cheesecakeutilities.Util.FirebaseUtils
 import com.itachi1706.cheesecakeutilities.Util.LogHelper
@@ -18,6 +28,7 @@ import java.util.*
  */
 abstract class BaseGpaFragment : Fragment(), SwipeEditDeleteCallback.ISwipeCallback {
     var callback: StateSwitchListener? = null
+    lateinit var adapter: GpaRecyclerAdapter
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
@@ -45,6 +56,30 @@ abstract class BaseGpaFragment : Fragment(), SwipeEditDeleteCallback.ISwipeCallb
         }
     }
 
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        val v = inflater.inflate(R.layout.fragment_recycler_view, container, false)
+        val recyclerView = v.findViewById<RecyclerView>(R.id.main_menu_recycler_view)
+
+        if (!evaluateToCont(v)) return v
+
+        recyclerView.setHasFixedSize(true)
+        val linearLayoutManager = LinearLayoutManager(context)
+        linearLayoutManager.orientation = LinearLayoutManager.VERTICAL
+        recyclerView.layoutManager = linearLayoutManager
+        recyclerView.itemAnimator = DefaultItemAnimator()
+        val itemTouchHelper = ItemTouchHelper(SwipeEditDeleteCallback(this, v.context))
+        itemTouchHelper.attachToRecyclerView(recyclerView)
+
+        // Update layout
+        adapter = GpaRecyclerAdapter(arrayListOf(), false)
+        recyclerView.adapter = adapter
+
+        callback?.onStateSwitch(getState())
+        return v
+    }
+
+    abstract fun getState(): Int
+    abstract fun evaluateToCont(v: View): Boolean
     abstract fun getLogTag(): String
     abstract fun initContextSelectMode(position: Int): Boolean
